requirements.txt
alembic==1.7.7
astroid==2.11.5
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
dill==0.3.5.1
Flask==2.1.1
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.7.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pylint==2.13.9
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.36
SQLAlchemy-Utils==0.38.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
